+++ include/nuttx/sensors/bmp180.h
@@ -68,7 +68,8 @@ extern "C"
  *   Register the BMP180 character device as 'devpath'
  *
  * Input Parameters:
- *   devpath - The full path to the driver to register. E.g., "/dev/press0"
+ *   path/no - The full path (or number) to the driver to register.
+ *             E.g.:("/dev/press0", i2c) or (0, i2c)
  *   i2c     - An instance of the I2C interface to use to communicate with
  *             BMP180
  *
@@ -76,8 +77,11 @@ extern "C"
  *   Zero (OK) on success; a negated errno value on failure.
  *
  ****************************************************************************/
-
+#ifndef CONFIG_SENSORS_BMP180_UORB
 int bmp180_register(FAR const char *devpath, FAR struct i2c_master_s *i2c);
+#else
+int bmp180_register_uorb(int devno, FAR struct i2c_master_s *i2c);
+#endif /* CONFIG_SENSORS_BMP180_UORB */
 
 #undef EXTERN
 #ifdef __cplusplus
